+++ shared/lib/transactions.utils.ts
@@ -1,4 +1,8 @@
-import { NestedTransactionMetadata } from '@metamask/transaction-controller';
+import {
+  NestedTransactionMetadata,
+  TransactionMeta,
+  TransactionType,
+} from '@metamask/transaction-controller';
 
 /**
  * Determines if a transaction is a batch transaction.
@@ -10,4 +14,28 @@ export function isBatchTransaction(
   nestedTransactions: NestedTransactionMetadata[] | undefined,
 ): boolean {
   return Boolean(nestedTransactions?.length);
+}
+
+/**
+ * Checks if a transaction (including nested transactions) matches any of the given types.
+ *
+ * @param transactionMeta - The transaction metadata to check.
+ * @param types - The transaction types to match against.
+ * @returns Whether the transaction or any nested transaction matches one of the types.
+ */
+export function hasTransactionType(
+  transactionMeta: TransactionMeta | undefined,
+  types: TransactionType[],
+) {
+  const { nestedTransactions, type } = transactionMeta ?? {};
+
+  if (types.includes(type as TransactionType)) {
+    return true;
+  }
+
+  return (
+    nestedTransactions?.some((tx) =>
+      types.includes(tx.type as TransactionType),
+    ) ?? false
+  );
 }
